requirements.txt
absl-py
accelerate==0.33.0
bitsandbytes==0.43.1
datasets==3.0.0
evaluate==0.4.3
huggingface-hub==0.23.5
immutabledict
langdetect
loguru==0.7.2
nltk
openai==1.48.0
openai-whisper==20231117
pandas==2.2.3
peft==0.11.1
protobuf==3.20.0
qa_metrics==0.2.17
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.14.1
sentencepiece==0.1.99
torchmetrics==1.4.2
tqdm==4.66.5
transformers @ git+https://github.com/huggingface/transformers@e15687fffe5c9d20598a19aeab721ae0a7580f8a
triton==2.1.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.12.1
librosa==0.10.2.post1



